MND-498 Add tests to ContributionWidget component
