+++ tests/singleProject.spec.js
@@ -33,38 +33,6 @@ describe('CaseStudies component', () => {
     expect(wrapper.element).toMatchSnapshot();
   });
 
-  test('shoud add contribution-widget__content_mobile-white class if projectName prop equals to guardrails', () => {
-    props = {
-      project: {
-        ...props.project,
-        projectName: 'guardrails'
-      }
-    };
-
-    wrapper = mount(SingleProject, {
-      propsData: props
-    });
-
-    const widget = wrapper.findAll('.contribution-widget__content-mobile').at(0);
-    expect(widget.classes()).toContain('contribution-widget__content_mobile-white');
-  });
-
-  test('shoud not add contribution-widget__content_mobile-white class if projectName prop notnequals to guardrails', () => {
-    props = {
-      project: {
-        ...props.project,
-        projectName: 'teacherly'
-      }
-    };
-
-    wrapper = mount(SingleProject, {
-      propsData: props
-    });
-
-    const widget = wrapper.findAll('.contribution-widget__content-mobile').at(0);
-    expect(widget.classes()).not.toContain('contribution-widget__content_mobile-white');
-  });
-
   test('.single-project__container shoud contain single-project__container_white-letters-theme class if isWhiteColored computed value will be true', () => {
     props = {
       project: {
@@ -112,29 +80,6 @@ describe('CaseStudies component', () => {
     expect(wrapper.props().project).toStrictEqual(props.project);
   });
 
-  test('shoud contain contribution-widget__img_red class if color will be #ec1c24', () => {
-    const widget = wrapper.findAll('.contribution-widget__img').at(0);
-    expect(widget.classes()).toContain('contribution-widget__img_red');
-    expect(widget.classes()).not.toContain('contribution-widget__img_black');
-  });
-
-  test('shoud contain contribution-widget__img_black class if color will be #000', () => {
-    props = {
-      project: {
-        ...props.project,
-        contributionWidgetColors: ['#000', '#000']
-      }
-    };
-
-    wrapper = mount(SingleProject, {
-      propsData: props
-    });
-
-    const widget = wrapper.findAll('.contribution-widget__img').at(0);
-    expect(widget.classes()).not.toContain('contribution-widget__img_red');
-    expect(widget.classes()).toContain('contribution-widget__img_black');
-  });
-
   test('shoud add single-project__background_first-project class to .single-project__background if projectName will be nambafood', () => {
     props = {
       project: {
@@ -164,23 +109,5 @@ describe('CaseStudies component', () => {
 
     const background = wrapper.findAll('.single-project__background').at(0);
     expect(background.classes()).not.toContain('single-project__background_first-project');
-  });
-
-  test('shoud contain contribution-widget__img_gray class if color will not be #000 or #ec1c24', () => {
-    props = {
-      project: {
-        ...props.project,
-        contributionWidgetColors: ['#fff', '#ccc']
-      }
-    };
-
-    wrapper = mount(SingleProject, {
-      propsData: props
-    });
-
-    const widget = wrapper.findAll('.contribution-widget__img').at(0);
-    expect(widget.classes()).not.toContain('contribution-widget__img_red');
-    expect(widget.classes()).not.toContain('contribution-widget__img_black');
-    expect(widget.classes()).toContain('contribution-widget__img_gray');
   });
 });
